Teach the program to recognize the ‘-v’ and ‘--version’ options
This patch allows the program to accept those two options, causing it
to print its version number and then exit successfully.  Trying to use
any other options makes the program exit with an error; the status
code for this error is one, and the previous exit code for failing to
find the remote repository now has the value two.  If the user
provides no options then the program lists all of the fork URLs
exactly like before.