Accept and document values for the ‘--format’ option
This patch is the first step towards displaying the output URLs in
different formats.  We begin by teaching the program to recognize the
‘-f’ and ‘--format’ options, requiring a value.  The patch lists the
acceptable values and provides an example URL of each format.
However, we currently do nothing with the new $FORMAT variable,
meaning it has no effect for now.