Stop requesting the beta version of GitHub’s JSON media type
The documentation at

    http://developer.github.com/v3/media/

explains the various media types which GitHub accepts.  We start using
the ‘application/vnd.github+json’ media type because GitHub is
planning to drop the beta versions sometime this year.  Our new media
type will always give us data in the latest format.  In the future we
may need to specify an API version but for the sake of future changes
we will structure the program around newest JSON data available.