Allow showing the owner of each fork
This patch adds two new command-line options: ‘-n’ and ‘--name’.
Providing either one of these options makes the program display a
username next to each URL showing who owns that fork.  We want the URL
and name to appear on the same line for use with ‘read’ in the future,
which means we have to use the ‘@text’ directive in our JSON query.
This lets avoid the newline, but requires we wrap all parts of the URL
in parentheses.

Implementing the feature demonstrated a problem with the while-loop
that processes the command-line options: the program would not always
read all options.  This patch corrects that behavior.