Implement the ‘-s’ and ‘--sort’ options for sorting forks
This option accepts the sorting values that the GitHub API accepts
when listing forks, documented at:

    http://developer.github.com/v3/repos/forks/

Currently we only accept values that we can pass directly to GitHub’s
own API.  It is technically possible for us to perform further sorting
based on values in the JSON data we receive; we can use the ‘jq’
program to perform such sorts.  But this patch does not attempt to add
that behavior, instead saving it as a feature to implement later.