Save the response from GitHub in a temporary file
This patch makes the program create a temporary file via ‘mktemp’,
which we document as a requirement.  Inside of that file we store a
copy of everything we receive from GitHub’s API, i.e. everything the
file has a copy of everything the ‘jq’ program receives.  We intend to
use this for debugging purposes only and therefore we do not tell the
user about the file.  Developers who need to debug the program and
want to examine that data can browse to the directory with the
temporary files and sort them by the most recent modification time.