Remove misuse of double-colons in the option definitions
Using a double-colon with getopt indicates that an option accepts an
optional argument.  The definitions for ‘--help’ and ‘--verbose’
incorrectly use the double-colon.  They accept no arguments.
Providing optional arguments to them causes no errors, but it is
semantically incorrect.  So this patch corrects the option definitions
to make it clear that neither ‘--help’ nor ‘--verbose’ accept any
command-line arguments.