README: Improve the notes about creating the manual
First this patch removes mention of the ‘--help’ option since Git
shows information about the alias, and not the script, when we try to
invoke it via that alias.  Second, this patch shows the Pandoc command
that will create the man page if the user does not have Tup installed
to perform that task automatically.