Remove ‘-h’ and ‘-v’, and replace ‘--help’ with ‘--usage’
We get rid of the short-hand notations for displaying help and version
information because we may want to use those for something else in the
future, e.g. ‘-v’ as short-hand for ‘--verbose’.

Running this program via an alias results in Git showing the alias
definition when given the ‘--help’ option.  This patch replaces that
option with ‘--usage’ to avoid that conflict.